shamap: refactor — deduplicate, extract shared utilities, remove dead code
- Extract keypath utilities into shamap/keypath.go (branch/nibble ops)
- Extract unified leaf node into shamap/leaf.go with leafKind dispatch
- Extract traversal primitives into shamap/traverse.go (WalkLeaves, WalkPreOrder, walkToKey)
- Deduplicate putItemUnsafe → delegates to putItemWithNodeTypeUnsafe
- Unify compare.go traversal: single walkBranch, handleLeafComparison, handleInnerComparison (no WithChannel variants)
- Delete dead code: cache.go, lru.go, memory_family.go, NewNodeID/FromBytes, ItemFromLeafNode, CachingSyncFilter/SyncState, VerifyProofPathDetailed
- Restore NewPebbleNodeStoreFamily (was incorrectly removed)
- Fix walkBranchCompare nil vs empty slice semantics in handleInnerComparison

~220 insertions, ~2,930 deletions across 23 files.
