+++ shamap/iterator.go
@@ -385,76 +385,6 @@ func (sm *SHAMap) LowerBound(id [32]byte) *Iterator {
 	return it
 }
 
-// firstBelow returns the first (smallest key) leaf node at or below the given node
-func (sm *SHAMap) firstBelow(node Node, parentID NodeID, branch int) LeafNode {
-	if node.IsLeaf() {
-		if leaf, ok := node.(LeafNode); ok {
-			return leaf
-		}
-		return nil
-	}
-
-	inner, ok := node.(*InnerNode)
-	if !ok {
-		return nil
-	}
-
-	nodeID, err := parentID.ChildNodeID(uint8(branch))
-	if err != nil {
-		return nil
-	}
-
-	for i := range BranchFactor {
-		child, err := sm.descend(inner, i)
-		if err != nil {
-			return nil
-		}
-		if child != nil {
-			result := sm.firstBelow(child, nodeID, i)
-			if result != nil {
-				return result
-			}
-		}
-	}
-
-	return nil
-}
-
-// lastBelow returns the last (largest key) leaf node at or below the given node
-func (sm *SHAMap) lastBelow(node Node, parentID NodeID, branch int) LeafNode {
-	if node.IsLeaf() {
-		if leaf, ok := node.(LeafNode); ok {
-			return leaf
-		}
-		return nil
-	}
-
-	inner, ok := node.(*InnerNode)
-	if !ok {
-		return nil
-	}
-
-	nodeID, err := parentID.ChildNodeID(uint8(branch))
-	if err != nil {
-		return nil
-	}
-
-	for i := BranchFactor - 1; i >= 0; i-- {
-		child, err := sm.descend(inner, i)
-		if err != nil {
-			return nil
-		}
-		if child != nil {
-			result := sm.lastBelow(child, nodeID, i)
-			if result != nil {
-				return result
-			}
-		}
-	}
-
-	return nil
-}
-
 // compareKeys compares two 32-byte keys lexicographically.
 // Returns -1 if a < b, 0 if a == b, 1 if a > b.
 func compareKeys(a, b [32]byte) int {
